+++ packages/react/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Next
 
+* Add `<Prefetch>` component for prefetching a route's data. Requires `@curi/router-prefetch`.
 * Forward `<Link>` ref to actual component (e.g. `<a>`).
 
 ## 1.0.0-beta.23